# openapi/payments.yaml
openapi: 3.1.0
info:
  title: Tap Payments API (Ethiopia-ready backend)
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/payments/charges:
    post:
      summary: Create a payment charge (Tap create-charge)
      operationId: createCharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChargeRequest"
      responses:
        "200":
          description: Charge created; return redirect URL to complete payment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateChargeResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Provider error while creating charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/payments/webhooks/tap:
    post:
      summary: Tap webhook endpoint (hashstring verification + idempotent state update)
      operationId: tapWebhook
      parameters:
        - in: header
          name: hashstring
          required: true
          schema:
            type: string
          description: Tap webhook signature header used to verify the payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TapChargeWebhookPayload"
      responses:
        "200":
          description: Webhook processed or recognized as duplicate.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/WebhookOkResponse"
                  - $ref: "#/components/schemas/WebhookDuplicateResponse"
        "401":
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/payments/{paymentId}:
    get:
      summary: Fetch payment state by internal payment id
      operationId: getPayment
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPaymentResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/payments/{paymentId}/refund:
    post:
      summary: Refund a payment (admin)
      operationId: refundPayment
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
        - in: header
          name: X-Admin-API-Key
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Refund accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefundResponse"
        "401":
          description: Missing/invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Invalid payment state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Provider does not support refunds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/payments/webhooks/chapa:
    post:
      summary: Chapa webhook endpoint (HMAC signature + idempotent state update)
      operationId: chapaWebhook
      parameters:
        - in: header
          name: x-chapa-signature
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapaWebhookPayload"
      responses:
        "200":
          description: Webhook processed or duplicate
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/WebhookOkResponse"
                  - $ref: "#/components/schemas/WebhookDuplicateResponse"
        "401":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
    PhoneInput:
      type: object
      properties:
        countryCode:
          type: string
          description: Phone country calling code (digits only, e.g. 251)
          pattern: "^[0-9]{1,3}$"
        number:
          type: string
          description: Subscriber phone number (digits only, typical 6-15 chars)
          pattern: "^[0-9]{6,15}$"
      required: [countryCode, number]
    CustomerInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          $ref: "#/components/schemas/PhoneInput"
      required: [phone]
    MetadataMap:
      type: object
      additionalProperties:
        type: string
    CreateChargeRequest:
      type: object
      properties:
        orderId:
          type: string
          description: Your internal order/payment reference id
          minLength: 1
        amount:
          type: number
          description: Amount to charge (positive decimal)
          exclusiveMinimum: 0
        currency:
          type: string
          description: Three-letter ISO currency code (e.g. ETB, USD, SAR)
          pattern: "^[A-Z]{3}$"
        customer:
          $ref: "#/components/schemas/CustomerInput"
        metadata:
          $ref: "#/components/schemas/MetadataMap"
      required: [orderId, amount, currency, customer]
    CreateChargeResponse:
      type: object
      properties:
        paymentId:
          type: string
        provider:
          type: string
          description: Payment provider name (currently "tap")
        providerChargeId:
          type: string
          description: Tap charge id (chg_...)
        redirectUrl:
          type: string
          description: URL to redirect the payer to complete the payment
        status:
          type: string
          description: Provider charge status (e.g. INITIATED/CAPTURED/FAILED)
      required: [paymentId, provider, providerChargeId, redirectUrl, status]
    TapChargeWebhookPayload:
      type: object
      description: Partial shape used by this backend to verify Tap signature and update state.
      properties:
        id:
          type: string
        object:
          type: string
        status:
          type: string
        amount:
          type: number
        currency:
          type: string
        reference:
          type: object
          properties:
            gateway:
              type: string
            payment:
              type: string
          required: [gateway, payment]
        transaction:
          type: object
          properties:
            created:
              type: string
          required: [created]
      required: [id, object, status, amount, currency, reference, transaction]
    WebhookOkResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["ok"]
      required: [status]
    WebhookDuplicateResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["duplicate"]
      required: [status]
    GetPaymentResponse:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        provider:
          type: string
        providerPaymentId:
          type: string
        status:
          type: string
        redirectUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderId, provider, providerPaymentId, status, createdAt, updatedAt]
    RefundRequest:
      type: object
      properties:
        amount:
          type: number
          description: Optional partial refund amount; defaults to full order amount
        reason:
          type: string
    RefundResponse:
      type: object
      properties:
        paymentId:
          type: string
        provider:
          type: string
        providerRefundId:
          type: string
        status:
          type: string
      required: [paymentId, provider, providerRefundId, status]
    ChapaWebhookPayload:
      type: object
      properties:
        tx_ref:
          type: string
        status:
          type: string
        reference:
          type: string
        amount: {}
        currency:
          type: string
      required: [tx_ref, status]